+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 #### Added
 - local-completion plugin
+- local-bashrc plugin
 
 ### 0.1.0 - 2015-10-01
 #### Added